setup health check controller
Since we can have multiple health checks we have to wire
them up trough multibind. This is a way to inject a set of an
instance to the service, so each play module can configure
up there own health checks. In the test scope we choose to
just skip them an have an empty set checks.