Healthcheck (#254)
* add stop watch

A basic stop watch that return the elapsed time
since it has been created.

* fix typo in UserService

* add health check service

basic health check service that can be wired up in the play project.

* setup health check controller

Since we can have multiple health checks we have to wire
them up trough multibind. This is a way to inject a set of an
instance to the service, so each play module can configure
up there own health checks. In the test scope we choose to
just skip them an have an empty set checks.

* style changes after review

Some small changes to the code style after feedback
from on the PR. Mostly just the use of execution context
and how we import it, making the code a bit more compact.

* replace implicit EC with play default

* add zabbix executor

Execute health checks and write the result to a file in
the zabbix format. We still need to find out the timezone
the reader of the file expects.

* new environment variables in docker setup

* ensure file exist before we try to write to it

* wire up zabbix health check as eager singleton

We need to have it eager so we start the scheduler.

* catch only non fatal exceptions

* group docker config under musit configuration

* use routes from the controller

In the future we want to remove the prepended string in the path.

* refactor to use akka streams to write the zabbix file

* Making Materializer and ActorSystem implicitly available in ZabbixExecutorProvider.
Allowing for implicit passing to ZabbixExecutor constructor.