refactor: move database config to must_service module
All the play projects that have a database does have a dependency
to this module. To reduce the copy/paste of the config and make it
a bit more clear what kind of database we include I moved it to
separate files.
Some module have custom config, for instance numThread config,
they still need to specify it in the play project.