test: load test resource files from classpath
Intellij Idea doesn't read the sbt config when running tests
from the editor. All tests that's not depending on configuration
loaded from the application.test.conf does work without any
modification. This is a workaround that enables us to run all the tests
trough Idea in a way that doesn't conflict with how sbt setup the test
environment.