Run tests from intellij (#294)
* chore: remove copyright header

* test: load test resource files from classpath

Intellij Idea doesn't read the sbt config when running tests
from the editor. All tests that's not depending on configuration
loaded from the application.test.conf does work without any
modification. This is a workaround that enables us to run all the tests
trough Idea in a way that doesn't conflict with how sbt setup the test
environment.

* test: ensure not to load test resources that doesn't exists

It's only the play projects that set the resource files trough sbt.
Now we're loading the resources if they exists, something they should
not as long as it's not a play project.

* test: refactor to not use java api