feat: consolidate fields in SampleObject
Some fields was like updated by and updated date can be represented as
one object. To make it more clear I did merge them and some other
fields. The reason to do this for only this class is that the new fields
that will be added in the next commit extended the class to have more
then 22 fields. The play json parser used tuples and it goes
up to 22 fields when using the Json.format function. The alternative
is to hand write all the fields for the writer and parser, something
that we don't want to do if we don't have to.