MUSARK-1001 - Bugfix/zabbix file (#298)
* fix: write over all the content in the file if it exists

When the content we want to write is shorter than the content in
the existing file we're ending up with some unwanted content from
the current file after we have written the new content. This ensure
that we clear the file before we start writing new content to it.

Resolves MUSARK-1001

* chore: remove copyright headers

* fix: change host group

"musit-developer" was a temporary value. Halvor did give us the proper
value to use.