Release bugfix (#299)
* fix: write over all the content in the file if it exists

When the content we want to write is shorter than the content in
the existing file we're ending up with some unwanted content from
the current file after we have written the new content. This ensure
that we clear the file before we start writing new content to it.

Resolves MUSARK-1001

* fix: change host group

"musit-developer" was a temporary value. Halvor did give us the proper
value to use.

* fix: use max size of matrix and preferred image (#292)

The size hints to the data matrix generator is ignored
so we have to use the actual size from the matrix it
self. the issue was that the image was cut and made
invalid/unable to scan.

Fixes MUSARK-994