Feature/musark 891 loan (#293)
* add loan tables and dao

This is a copy from another branch after we decided to merge two
service project. Atm we have the event table defined twice since they're
not equal when it comes to the event type, but should be merged to one
table in the future.

* add missing table in migration script

* mode evolution script to test scope

* move loan to separate packages

This is to align the codebase to how it's done for analysis

* keep the loan data in a loan schema

* reformat code after changing to scalafmt

* refactor: use non deprecated code for slick

This branch is so old that it used the old api for slick.

* test: use MusitResultValues

* style: use case match in map to make the code more readable

Thank you codacy :-)

* feat: add museum it to loan event

* fix: use correct column name for lent object id

* feat: add external ref

* refactor: remove unused function

* refactor: rename domain classes

* refactor: move common code to controller package

* feat: query for active loan events

* feat: create and find endpoint for loan

- Create a loan endpoint
- Find all active loans

* feat: add required fields when creating a loan

* feat: add support for bad request trough MusitValidationError

* feat: do not accept loans without objects

* test: improve feedback from musit success values

This will include the stacktrace from the musit db error. For now
it's the only musit error that have a throwable included. In the
future we should probably have it as optional for all error

* style: fix typo in implicit value

* test: move loan db migration into 1.sql

* chore: add loan to oracle init script

This commit fixes an issue where the musit user wasn't granted the
rights to musark_analysis.

* refactor: change name for external ref to case number

It's a case number and not a general external ref. We will need a
external ref (not in this PR) but it will have both a value and a
source.

* refactor: remove usage of timestamp

* refactor: prefix tables with loan and analysis

Since we have two tables traits in this module it will be more clear
if they didn't have the same name.

* fix: parsing of dates and removed parser for get active loans

The test could probably have been written better but it's better then
nothing. It found some bugs at least :-)

* test: include case numbers in test.