feat: update index after a give timestamp
Now we can trigger a update of the index after a given timestamp. It
has stored the latest update stamp for each alias so we can easy
update the latest inserted and updated documents.

I'm not totally happy with where it abstractions is atm since we're
now having indexing and some maintaining for the index in the same
place. We need to take a closer look at the abstractions here.