refactor: make flow definitions/pipeline for each indexed type
The custom work will be done in the flow pipeline so it will make it
easier to customize them when they're separated from each other.
Also the integration with elasticsearch is re written and abstracted
away form the indexed typed flows.

Still some small reporting back to the IndexActor that need to be done
but now we have at least a better abstraction.