refactor: chain the two steps in a separate function to make it more clear
We have at least two steps when streaming the documents. The first step
is to populate/convert the documents from it's source to the format we
want to index them. Then we need to make it into an bulk action. This
need to be done in the TypedFlow step since we have better knowledge
on how the document should be indexed. For instance if it should have
a parent-child relation.