feat: include parent object into the child object response from es
When returning samples we need to include the musit object as
inner-hit. This feature is not working in the search query definition
for the current version of elastic4s client. The pull requests that
fixes the issues are accepted so we're waiting for a new release
(will be available in version >= 5.4.12)